+++ pages/brochure-2.js
@@ -1,0 +1,10 @@
+import Layout from 'components/layout/webview-layout'
+import BrochureScreen from 'screens/brochure'
+
+export default function Brochure() {
+    return (
+        <Layout title={'Brochure Hemat'}>
+            <BrochureScreen title="Brochure Hemat" hemat />
+        </Layout>
+    )
+}